Struct stm32f0xx_hal::adc::Adc

pub struct Adc { /* fields omitted */ }

Analog to Digital converter interface

Implementations

impl Adc

pub fn new(adc: ADC, rcc: &mut Rcc) -> Self

Init a new Adc

Sets all configurable parameters to defaults, enables the HSI14 clock for the ADC if it is not already enabled and performs a boot time calibration. As such this method may take an appreciable time to run.

pub fn save_cfg(&mut self) -> StoredConfig

Saves a copy of the current ADC config

pub fn restore_cfg(&mut self, cfg: StoredConfig)

Restores a stored config

pub fn default_cfg(&mut self) -> StoredConfig

Resets the ADC config to default, returning the existing config as a stored config.

pub fn set_sample_time(&mut self, t_samp: AdcSampleTime)

Set the Adc sampling time

Options can be found in AdcSampleTime.

pub fn set_align(&mut self, align: AdcAlign)

Set the Adc result alignment

Options can be found in AdcAlign.

pub fn set_precision(&mut self, precision: AdcPrecision)

Set the Adc precision

Options can be found in AdcPrecision.

pub fn max_sample(&self) -> u16

Returns the largest possible sample value for the current settings

pub fn read_abs_mv<PIN: Channel<Adc, ID = u8>>(&mut self, pin: &mut PIN) -> u16

Read the value of a channel and converts the result to milli-volts