Struct stm32f0xx_hal::gpio::gpiod::PD8

pub struct PD8<MODE> { /* fields omitted */ }

Pin

Implementations

impl<MODE> PD8<MODE>

pub fn into_alternate_af0(self, _cs: &CriticalSection) -> PD8<Alternate<AF0>>

Configures the pin to operate in AF0 mode

pub fn into_alternate_af1(self, _cs: &CriticalSection) -> PD8<Alternate<AF1>>

Configures the pin to operate in AF1 mode

pub fn into_alternate_af2(self, _cs: &CriticalSection) -> PD8<Alternate<AF2>>

Configures the pin to operate in AF2 mode

pub fn into_alternate_af3(self, _cs: &CriticalSection) -> PD8<Alternate<AF3>>

Configures the pin to operate in AF3 mode

pub fn into_alternate_af4(self, _cs: &CriticalSection) -> PD8<Alternate<AF4>>

Configures the pin to operate in AF4 mode

pub fn into_alternate_af5(self, _cs: &CriticalSection) -> PD8<Alternate<AF5>>

Configures the pin to operate in AF5 mode

pub fn into_alternate_af6(self, _cs: &CriticalSection) -> PD8<Alternate<AF6>>

Configures the pin to operate in AF6 mode

pub fn into_alternate_af7(self, _cs: &CriticalSection) -> PD8<Alternate<AF7>>

Configures the pin to operate in AF7 mode

pub fn into_floating_input(self, _cs: &CriticalSection) -> PD8<Input<Floating>>

Configures the pin to operate as a floating input pin

pub fn into_pull_down_input(self, _cs: &CriticalSection) -> PD8<Input<PullDown>>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self, _cs: &CriticalSection) -> PD8<Input<PullUp>>

Configures the pin to operate as a pulled up input pin

pub fn into_analog(self, _cs: &CriticalSection) -> PD8<Analog>

Configures the pin to operate as an analog pin

pub fn into_open_drain_output(
self,
_cs: &CriticalSection
) -> PD8<Output<OpenDrain>>

Configures the pin to operate as an open drain output pin

pub fn into_push_pull_output(
self,
_cs: &CriticalSection
) -> PD8<Output<PushPull>>

Configures the pin to operate as an push pull output pin

pub fn into_push_pull_output_hs(
self,
_cs: &CriticalSection
) -> PD8<Output<PushPull>>

Configures the pin to operate as an push pull output pin with quick fall and rise times

impl PD8<Output<OpenDrain>>

pub fn internal_pull_up(&mut self, _cs: &CriticalSection, on: bool)

Enables / disables the internal pull up

impl<AF> PD8<Alternate<AF>>

pub fn internal_pull_up(self, _cs: &CriticalSection, on: bool) -> Self

Enables / disables the internal pull up

impl<AF> PD8<Alternate<AF>>

pub fn set_open_drain(self, _cs: &CriticalSection) -> Self

Turns pin alternate configuration pin into open drain

impl<MODE> PD8<Output<MODE>>

pub fn downgrade(self) -> Pin<Output<MODE>>

Erases the pin number from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type

impl<MODE> PD8<Input<MODE>>

pub fn downgrade(self) -> Pin<Input<MODE>>

Erases the pin number from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type