Trait stm32f0xx_hal::prelude::_embedded_hal_gpio_StatefulOutputPin

pub trait _embedded_hal_gpio_StatefulOutputPin: OutputPin {
    fn is_set_high(&self) -> Result<bool, Self::Error>;

    fn is_set_low(&self) -> Result<bool, Self::Error>;
}

Push-pull output pin that can read its output state

This trait is available if embedded-hal is built with the "unproven" feature.

Required methods

fn is_set_high(&self) -> Result<bool, Self::Error>

Is the pin in drive high mode?

NOTE this does not read the electrical state of the pin

fn is_set_low(&self) -> Result<bool, Self::Error>

Is the pin in drive low mode?

NOTE this does not read the electrical state of the pin

Loading content...